Support compiling with features that are not supported in DEX.
This change refactors the errors for unsupported features to non-final
errors. The --map-diagnostics flag/API can be used to make these
warnings instead. In so doing, the compiler will replace the unsupported
instructions by unconditional throwing instructions with a descriptive
error message.

This CL also adds additional diagnostics types for each of the
unsupported features as part of the future compiler API.

Bug: b/174733673
Bug: b/154778581
Bug: b/198142625
Change-Id: I05d7965d7f9d49b97a5ac06bce301d7a861be802