Limit the number of possible equivalence groups.
This CL hardcodes a limit on the number of actual groups for a given set
of potential equivalences to ensure that the quadratic behavior is not
hit. The expected groups should generally be small for a given type of
synthetics so this will not likely have an impact on size.

It would be beneficial to follow-up with a sorting-based implementation
and avoid the arbitrary limit.

Bug: b/237413146
Bug: b/236690153
Change-Id: Id4e57e83a5a0b1431e95753d605c52865c4b05b9