Add conditional and non-startup classification to startup descriptors
This updates generate_startup_descriptors.py to identify the logcat message "Displayed <app id>/..." from ActivityTaskManager. All startup descriptors emitted after this message are classified as "post startup" and will only be included in the generated startup list when running with --include-post-startup.

This also classifies startup descriptors that are only seen in some but not all launches of the app as "conditional" startup descriptors. As an example, a background task that may or may not execute fully during startup could be classified as conditional. Conditional startup descriptors are only included in the generated startup list if running with --include-conditional-startup.

Change-Id: I46d6d07748dc863a8df44ca59dbf90d0ea34dcef