Log synthetic context from synthetics in startup instrumentation
Prior to this CL, when a synthetic method was executed, the instrumentation would print the reference of the synthetic method, where the holder is replaced by the synthetic context. This gives more knowledge of which synthetic methods are executed.

Example trace event: "SLcom/example/SyntheticContext;->syntheticMethod()V"

For simplicity, this changes the instrumentation of synthetic methods to only print the synthetic context. Follow up work is then needed to improve the precision for synthetics.

New trace event: "SLcom/example/SyntheticContext;"

The interpretation of the new trace event is that any method of any synthetic from com.example.SyntheticContext has been executed.

Change-Id: Ic76ef7b879a582e5f2d6b5555e4d51da690b6391