Insert new startup descriptors at point of execution
When generating a startup list using generate_startup_descriptors.py, the app is launched multiple times until the set of startup descriptors no longer change.

Prior to this CL, newly identified startup descriptors from launching the app were simply appended to the existing startup list.

For a new startup descriptor from a newly generated startup list, this CL finds the closest startup descriptor that is already in the complete startup list, and then inserts the new startup descriptor immediately after that in the existing trace (instead of appending it to the end).

Change-Id: Iee38a80e5c2b15ddc207a704204e1ef4e942c0c6