New compiler API for passing startup profile to D8/R8
This adds a new setStartupProfileProvider(StartupProfileProvider) API to D8 and R8 command builders (D8Command$Builder and R8Command$Builder).

StartupProfileProvider is an interface with a single get() method that returns the startup profile as a java.lang.String, and is added to the public API of the compiler.

The startup profile is used by D8 and R8 to ensure that classes used during startup will be present in the primary classes.dex file. Moreover, R8 uses the startup profile to guide its optimizations so that (additional) post-startup code is not added to the app's startup classes.

Change-Id: Ic114a3de785219508a2c634a297f3cef226d5bbf