Introduce preliminary API classes for accepting an ART profile
This introduces a set of API interfaces for providing an ART profile to the compiler and a consumer that accepts the rewritten ART profile at the end of compilation.

The implementation of this API will provide an initial, incomplete implementation for b/237043695, which should be evaluated against the existing rewriting of ART baseline profiles using a mapping file.

Bug: b/237043695
Change-Id: Iecdd1cede9163a9ca8a0f48a91479b113c56900a