[Compose] Allow for lossy composition
This change also moves the residual signature out of the MemberMapping and MappedRange objects unless we could read the residual signature. The reason is that the residual mapping has previously been defined as the orignal signature with a renamed name and therefore we cannot use it for composition purposes.

In principle, every lookup on residual signatures should go through the computation but there is a lot of internal references relying on reading a residual signature. These should be removed and the cleanup is tracked in b/246919706.

Finally, all tests are bumped to use map version experimental such that we can rely on stable right hand sides.

Bug: b/241763080
Bug: b/246919706
Change-Id: Ib86450eb03e3ccea25cb31ab738692570e8308e2