Do not enable filled-new-array for pre-N minSdkVersion
ART 6.0.1 tests show an odd failure with them, so disable for pre-N.
Many apps now have N as their minSdkVersion, so this won't be a missed
opportunity for most (or at least this will be the case soon).

This still enables String[] for pre-N so as to not regress the status
quo.

Bug: b/246971330
Change-Id: I0a87f6dc8201b487bce0ff2af0abd35f32fd03a8