Add tests for conditional keep rules with classmember rule based on subtypes
This will test that we can do a conditional keep rule on the form:
-if class * extends Foo
  -keepclassmembers class * {
    <1> *;
  }

Where this will keep fields if the type of the field extends Foo (and the class is live)

This also includes testing of extends on the conditional being the same type.
In the case above, the rule will NOT keep a field on a class that have the field typed as Foo.
If this is needed (which it is for b/264686688) one would need to do:
-keepclassmembers class * {
  Foo *;
}
-if class * extends Foo
  -keepclassmembers class * {
    <1> *;
  }

Bug: 264686688
Change-Id: I957fe4ba0387b114553025aa4dd11af980658ecc