Fixes to writing the DEX conatiner format
 * Sort the map by offset
 * Calculate signature and checksum correctly

Before the entries in the map was written as for non-container DEX,
but as the string section can now be after the other sections (as it
is shared with a later section) the map needs to be sorted after the
location of the shared string table is known.

Changing the map also affects the signature and checksum, so these
has to be calculated after all offsets has been finalized.

Bug: b/249922554
Change-Id: I1c7e8ac8c735c03cc70d88ff6824f16af7f72625