[KeepAnno] Generalize the mapping to compatible rules.
This introduces a target kind that distinguishes between targets that
should include the class or member targets as a disjunction (e.g.,
keep), or as a conjuction (e.g., keepclasseswithmembers) or if the
targets are only the members (e.g., keepclassmembers).

The rule printing now avoids splitting the class and members rules in
the disjunctive case when possible. When a class target is needed, its
rule is amended with the Object::finalize member to avoid keeping
<init>().

Bug: b/248408342
Change-Id: I6cd024ec362f4e23d91778e405a5dcebacaafb87