Wrapper around desugaring event consumers for profile rewriting
This CL rewrites the uses of {D8,R8}CfInstructionDesugaringEventConsumer to program up against the CfInstructionDesugaringEventConsumer.

This enables the introduction of a wrapper around these two, which can augment the ART baseline profile (if any) when desugaring happens.

The wrapper introduced by this CL currently only forwards to the {D8,R8}CfInstructionDesugaringEventConsumer, and is not currently being attached (i.e., it is still effectively unused).

Bug: b/265729283
Change-Id: Icfdc24e458604857762febc2f6cd1cb2fa62a5a0