Remove support for -overloadaggressively
Fixes: b/221174453

RELNOTES: Removed support for keep flag -overloadaggressively.
The flag remains a valid flag and will be parsed by R8. If present, R8
will silently ignore it and continue compilation as if it was not set.
Previously compiling with -overloadaggressively would result in larger
output size as more line information is needed to ensure that stacks can
be retraced correctly.

Change-Id: Idd4f5fd4afb6cca21dd2845e5d6573c197613c5c