Extend profile rewriting to nest based access desugaring
Since the synthetics from nest based access desugaring are created in the prepare desugaring step, this CL extends the prepare() step with the CfInstructionDesugaringEventConsumer, and performs the relevant callbacks to it.

Additionally, this CL makes some minor cleanups to the nest based access desugaring. In particular this CL clearly separates the parts of nest based access desugaring that creates methods (the prepare step) from the parts that don't (the instruction rewrite step). As an example, prior to this CL, the method getMethodBridgeReference() had logic to create a synthetic constructor argument class using synthetic items. Moreover, D8 specific handling of classpath accesses to the program is moved to D8NestBasedAccessDesugaring.

Bug: b/265729283
Change-Id: I6f8da35fffad467793a9fa5f744204b252c06e8f