Account for nested synthetics in profile rewriting
This adds support for nested synthetics in the profile rewriting by checking if the synthesizing context is in the art profile *or* in the current set of additions to the art profile.

When the context is a method in the art profile, we transfer its flags to the synthetic method when we add it to the profile.

When the context is a method in the current set of additions to the art profile, we must wait transferring the flags to the synthetic method until commit time, since its flags may change if the context has multiple synthesizing contexts.

Example:

If context m3() has synthesizing contexts m1() and m2(), and m1() is 'HOT' and m2() is 'STARTUP', then m3() will be marked as 'HOT' after processing m1(). If m4() is synthesized from m3() before processing m2(), then the flags of m4() will be 'HOT'. To ensure m4() also receives the 'STARTUP' flag, the dependencies are recorded in a graph so that we can propagate the flags through-out the edges of the graph before committing the additions.

Bug: b/265729283
Change-Id: Ie50b3027ce942b877ad50785c4e5ccb08d74112d