[LIR] Abstract encoding and decoding of values to distinguish phis.
This CL does not change the actual encoding or change the identification
of phis. It introduces most of the infrastructure to make the
identification possible independent of the base writer and
parser/iterator. Follow-up CLs will create a new strategy and amend the
builders to allow a separate value-table for phis.

Bug: b/225838009
Change-Id: I22fdb736a3e6c8d3bc4a249ec21d982b3ea2aea8