Write the Record attribute for both CF and DEX
This adds representation of all the information from the Record
attribute to the compiler.

Before the Record attribute was implicitly generated from the instance
fields of the record class. However, the Record attribute carry
this information separately in the class file where the annotations on the
field and the record component can be different based on the @Target
(@Target(ElementType.FIELD) vs. @Target(ElementType.RECORD_COMPONENT))
used when declaring the annotation used to annotate components of a record.

When writing to DEX a system annotation of type dalvik.annotation.Record
encodes the Record attribute information.

The shrinking and minification handling in R8 is not complete.

Bug: b/231930852
Bug: b/274888318
Change-Id: Ib9dd1876bca84e6aa735141b094d77f4a57de315