Don't cannonicalize code objects when reading
Previously we would not canonicalize code objects, but our reader
would still assume that the object could be canonicalized and would
leave them canonicalized if they where.

The remainder of the pipeline assumes that we can change the code
objects, e.g., when jumbostring rewriting where we hit this when
merging already dexed and dedupped code objects.

Bug: 283145010
Change-Id: I409fe3fbcc56a87b604cc6c0ae2021b7d8e0c088