Refactor lenses to describe their one-step mapping only
We previously introduced the methods getPreviousMethodSignature() and getNextMethodSignature(), which describe the mapping made by a single lens. These are used to implement getOriginalMethodSignature() and getRenamedMethodSignature(), so that each lens does not need to add an implementation that recurses into the previous lens.

This CL finishes this cleanup by introducing the methods getPreviousClassType(), getNextClassType(), getPreviousFieldSignature() and getNextFieldSignature(), and using these to implement getOriginalClassType(), lookupType(), getOriginalFieldSignature() and getRenamedFieldSignature().

Change-Id: I1e4f6df8feb08a52d979572d453735a610a626e5