Towards using a new abstraction for definitely set/unset bits
This cleans up the joining of abstract values by introducing a new AbstractValueJoiner class that encapsulates all logic related to joining (including which and when to use specific abstractions).

This also introduces a new DefiniteBitsNumberValue abstraction that can be used for integer values. Changing the join of abstract values to use this new abstraction is left for future work.

Bug: b/196017578
Fixes: b/196321452
Change-Id: I5fba6fdc8d04c185d9b932c08b354c0eff5f97a2