Add skeleton for resource shrinker integration
This adds a new interface for tracing through the resource model.

In the old gradle setup this is implemented just as a no-op, with copy through.
For the new gradle setup this is implemnted with a single call to the resource shrinker to ensure access, and then just copy through.

Bug: b/287398085
Change-Id: I9e103f1149c9e209ab7dbad607a5047aac83cfbd