Extend methods relying on abstract value with abstract value supplier
This introduces a new helper for computing an abstract value from a given value.

This is normally done using Value#getAbstractValue. However, the result of this method may rely on the abstract values of some of the in values to its definition (e.g., v2 = v0 + v1).

In order to avoid that Value#getAbstractValue computes a "deep" abstract value by evaluating large arithmetic expressions, this CL introduces an abstract value supplier that implements the computation of the abstract value.

This is used to implement a "shallow" abstract value that uses Value#getAbstractValue for the first call, but AbstractValue#unknown for all recursive calls.

Change-Id: Iba941d3d11ec9881848d77555cc21e0b50ed3630