Allow generation of resources based on test R classes
This allows you to write R classes directly in tests and access the fields from other classes.

The R classes are used to populate the resources with default values
(the name for strings, a small png for drawables)

The non R classes are rewritten to point at the right R class fields.
We need this to avoid having the outer class.

Bug: b/287398085
Change-Id: I3ecc883f4c6f4ebbc91aceb6b0db65e044a3dd2d