Insert pop in IR when a catch block has no use for the pushed exception value
Must be done after goto inlining, as to properly detect if predecessors have a try-catch relation to the target block.

Assumes the IR will push a MoveException into the first instruction of blocks when a catch block uses the exception value. When this is missing but the block is a catch, we will insert our pop.