Test local variable table for parameter with signature
Parameters of types with generic signatures cannot be encoded in the
header of the debug_info_item, therefor the parameter is set to null
and the event stream will have a start local extended event active at
pc 0 that declares the parameter as active.

From DEX VM 7 and onward, the encoding results in the VM reporting an
empty range, i.e., [0:0[, for the parameter as well as the valid range
of the parameter with its signature. Some debugger environments do not
deal correctly with this unexpected empty range.

The current output of D8 is to emit null in the parameter list and
issue the START_LOCAL_EXTENDED event after the +0+0 DEFAULT event.

This test checks that the result is the same regardless of the order
of the START_LOCAL_EXTENDED event and the DEFAULT event.

It is not possible for the compiler to emit any other encoding of
parameters for types with generic signatures without resulting in
incorrect/incomplete behavior when stepping. Ensuring a non-empty
range for the initial parameter would cause the method entry to no
longer have the generic signature information.

Bug: b/297843934
Change-Id: I03633d43a330d79c5e5ee3ebe4086374d42ed0be