Per module testing state with shared index
This moved the creation of the shared state index to test.py such
that it is created once and ahead of test runs. Each test module
emits to the shared index and appends a status block on completion.
Test state is moved to per-module sub-directories and filters are
computed based on the state of the tests in each modules sub-dir.

Bug: b/297316723
Change-Id: Ie20754600125c3d53194bd684f7c976ca96d4926