Support rerunning all previously failing tests
This also moves the "outstanding" functionality to a "run mode".
It appears to often time out if the test state is large. A mode
to rerun all tests is also added and replaces the need to delete
the test state to start over.

The testing state functionality is removed from the old build.

Bug: b/297316723
Change-Id: I5e3e239c6409d1644305599e28b92d5b3730686c