Preserve positions for throwing instruction types
This avoids truncating positions over throwing instructions we cant then
be expanded out again. The truncation leads to instructions not covered
by the same potential inline frames which causes lack of precision in
mapping files. The cost of maintaining the position should be low to
none as the shared debug output encoding should eliminate the overhead.

Change-Id: Idc92e1d59ecc8113401bd97ee83f7b515f5da80b