Refactor single dispatch target to allow avoiding duplicate resolution
This refactors the implementation of `InvokeMethod#lookupSingleTarget` to `SingleResolutionResult#lookupDispatchTarget`.

This ensures that we have always performed successful resolution prior to looking up the dispatch target, and that the (single) resolution result is always available for the dispatch target lookup, meaning we can avoid duplicating resolution as done currently.

This CL is meant to be a pure refactoring, thus the change to avoid duplicating the resolution is postponed to a follow up.

Change-Id: I326f336865c1cf96d2db898a967fc83154b245dd