Initial setup for doing legacy resource shrinking in R8
With legacy we mean do exactly like the current pipeline, which is
first dexing the app, then running resource shrinking on the output
dex and the resources.

This tries to minimize the changes to the shrinker library outside
r8integration to ease sync with studio.

This also maintains compatability with the old version of android
common used internally.

Bug: 305892375
Bug: 287398085
Change-Id: I062aad33340d1d165f8220df6b710a8b829c2b2b