Introduce a threading module
This CL updates just the reader to use the new modules.
The blocking module is the default module and is bundled in the R8 deps
as well as in the relocated R8 release.

For R8 without dependencies one of the two threading modules must be
added as a dependency on the runtime classpath.

Bug: b/304992619
Change-Id: Id1f765b2b4f33f67bc623baab55e98608d39648a