Make addPossiblyThrowingInstructionsToPossiblyThrowingBlock() work in more cases
It did not work when the cursor was before a throwing instruction. It
would split the block, which would move the throwing instruction to the
subsequent block, and then add another throwing instruction to the
subsequent block.

This also adds an override that takes a Collection<> to avoid needing an
array.

Bug: b/244238384
Change-Id: I608e72cb5e1100355820141404d874ddee961cc6