Desugar records again for Android U (API level 34)
Art had a bug in native record support in the initial release of
Android U.

This will be fixed with Art mainline updates, but turning record
desugaring on again for API level 34 to avoid hitting issues on
Android U where mainline updates are not been applied.

Also removed the predicate canUseNativeRecords and replaced with
isRecordsDesugaredForD8 and isRecordsDesugaredForR8. These two
together with runtimeWithRecordsSupport are now used in all record
tests.

Change-Id: I111e9779b657492583cd7b9eb60788d8ec6d198e