Simplify dominator-related code in ValueUtils
Rather than have a helper to compute all dominators between sourceBlock
and destBlock, use a helper that checks only whether given blocks are
dominators (by seeing if a path exists from dest to source that does not
go through the block in question).

This is much simpler both in terms of code and worst-case runtime.

Bug: 244238384
Change-Id: I3e11e98a12551c04b51f95f4666239c3f8e7133d